requirements.txt
faiss-cpu==1.7.4
google-ai-generativelanguage==0.6.10
google-api-core==2.24.0
google-api-python-client==2.156.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.66.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
hyperlink==21.0.0
langchain==0.3.13
langchain-core==0.3.28
langchain-text-splitters==0.3.4
langsmith==0.2.6
numpy==1.26.4
pandas==2.2.3
PyPDF2==3.0.1
tensor==0.3.6
torch==2.5.1
python-dotenv==1.0.1